+++ sickrage/core/databases/main/index.py
@@ -23,6 +23,20 @@ from hashlib import md5
 from CodernityDB.hash_index import HashIndex
 
 
+class MainVersionIndex(HashIndex):
+    _version = 1
+
+    def __init__(self, *args, **kwargs):
+        kwargs['key_format'] = 'I'
+        super(MainVersionIndex, self).__init__(*args, **kwargs)
+
+    def make_key(self, key):
+        return key
+
+    def make_key_value(self, data):
+        if data.get('_t') == 'version' and data.get('database_version'):
+            return data.get('database_version'), None
+
 class MainTVShowsIndex(HashIndex):
     _version = 1
 
@@ -156,6 +170,7 @@ class MainWhitelistIndex(HashIndex):
     def make_key_value(self, data):
         if data.get('_t') == 'whitelist' and data.get('show_id'):
             return data.get('show_id'), None
+
 
 class MainFailedSnatchesIndex(HashIndex):
     _version = 1