+++ sickrage/core/searchers/backlog_searcher.py
@@ -134,8 +134,8 @@ class BacklogSearcher(object):
                     elif curQuality != Quality.UNKNOWN and curQuality > max(anyQualities):
                         continue
 
-            # only fetch if not archive on first match
-            if curStatus == DOWNLOADED and show.archive_firstmatch:
+            # skip upgrading quality of downloaded episodes if enabled
+            if curStatus == DOWNLOADED and show.skip_downloaded:
                 continue
 
             epObj = show.getEpisode(int(result["season"]), int(result["episode"]))