+++ sickrage/notifiers/emailnotify.py
@@ -64,11 +64,11 @@ class EmailNotifier(Notifiers):
                 try:
                     msg = MIMEMultipart('alternative')
                     msg.attach(MIMEText(
-                        "<body style='font-family:Helvetica, Arial, sans-serif;'><h3>SiCKRAGE Notification - Snatched</h3>\n<p>Show: <b>" + re.search(
-                            "(.+?) -.+", ep_name).group(1) + "</b></p>\n<p>Episode: <b>" + re.search(
-                            ".+ - (.+?-.+) -.+", ep_name).group(
-                            1) + "</b></p>\n\n<footer style='margin-top: 2.5em; padding: .7em 0; color: #777; border-top: #BBB solid 1px;'>Powered by SiCKRAGE.</footer></body>",
-                        'html'))
+                        "<body style='font-family:Helvetica, Arial, sans-serif;'><h3>SiCKRAGE Notification - "
+                        "Snatched</h3>\n<p>Show: <b>" + re.search("(.+?) -.+", ep_name).group(1) +
+                        "</b></p>\n<p>Episode: <b>" + re.search(".+ - (.+?-.+) -.+", ep_name).group(1) +
+                        "</b></p>\n\n<footer style='margin-top: 2.5em; padding: .7em 0; color: #777; border-top: #BBB "
+                        "solid 1px;'>Powered by SiCKRAGE.</footer></body>", 'html'))
                 except:
                     try:
                         msg = MIMEText(ep_name)
@@ -86,7 +86,7 @@ class EmailNotifier(Notifiers):
                                   msg):
                     sickrage.app.log.debug("Snatch notification sent to [%s] for '%s'" % (to, ep_name))
                 else:
-                    sickrage.app.log.error("Snatch notification ERROR: %s" % self.last_err)
+                    sickrage.app.log.warning("Snatch notification WARNING: %s" % self.last_err)
 
     def notify_download(self, ep_name, title="Completed:"):
         """
@@ -106,11 +106,11 @@ class EmailNotifier(Notifiers):
                 try:
                     msg = MIMEMultipart('alternative')
                     msg.attach(MIMEText(
-                        "<body style='font-family:Helvetica, Arial, sans-serif;'><h3>SiCKRAGE Notification - Downloaded</h3>\n<p>Show: <b>" + re.search(
-                            "(.+?) -.+", ep_name).group(1) + "</b></p>\n<p>Episode: <b>" + re.search(
-                            ".+ - (.+?-.+) -.+", ep_name).group(
-                            1) + "</b></p>\n\n<footer style='margin-top: 2.5em; padding: .7em 0; color: #777; border-top: #BBB solid 1px;'>Powered by SiCKRAGE.</footer></body>",
-                        'html'))
+                        "<body style='font-family:Helvetica, Arial, sans-serif;'><h3>SiCKRAGE Notification - "
+                        "Downloaded</h3>\n<p>Show: <b>" + re.search("(.+?) -.+", ep_name).group(1) +
+                        "</b></p>\n<p>Episode: <b>" + re.search(".+ - (.+?-.+) -.+", ep_name).group(1) +
+                        "</b></p>\n\n<footer style='margin-top: 2.5em; padding: .7em 0; color: #777; border-top: #BBB "
+                        "solid 1px;'>Powered by SiCKRAGE.</footer></body>", 'html'))
                 except:
                     try:
                         msg = MIMEText(ep_name)
@@ -128,7 +128,7 @@ class EmailNotifier(Notifiers):
                                   msg):
                     sickrage.app.log.debug("Download notification sent to [%s] for '%s'" % (to, ep_name))
                 else:
-                    sickrage.app.log.error("Download notification ERROR: %s" % self.last_err)
+                    sickrage.app.log.warning("Download notification WARNING: %s" % self.last_err)
 
     def notify_subtitle_download(self, ep_name, lang, title="Downloaded subtitle:"):
         """
@@ -148,11 +148,12 @@ class EmailNotifier(Notifiers):
                 try:
                     msg = MIMEMultipart('alternative')
                     msg.attach(MIMEText(
-                        "<body style='font-family:Helvetica, Arial, sans-serif;'><h3>SiCKRAGE Notification - Subtitle Downloaded</h3>\n<p>Show: <b>" + re.search(
-                            "(.+?) -.+", ep_name).group(1) + "</b></p>\n<p>Episode: <b>" + re.search(
-                            ".+ - (.+?-.+) -.+", ep_name).group(
-                            1) + "</b></p>\n<p>Language: <b>" + lang + "</b></p>\n\n<footer style='margin-top: 2.5em; padding: .7em 0; color: #777; border-top: #BBB solid 1px;'>Powered by SiCKRAGE.</footer></body>",
-                        'html'))
+                        "<body style='font-family:Helvetica, Arial, sans-serif;'><h3>SiCKRAGE Notification - Subtitle "
+                        "Downloaded</h3>\n<p>Show: <b>" + re.search("(.+?) -.+", ep_name).group(1) +
+                        "</b></p>\n<p>Episode: <b>" + re.search(".+ - (.+?-.+) -.+", ep_name).group(1) +
+                        "</b></p>\n<p>Language: <b>" + lang + "</b></p>\n\n<footer style='margin-top: 2.5em; padding: "
+                                                              ".7em 0; color: #777; border-top: #BBB solid "
+                                                              "1px;'>Powered by SiCKRAGE.</footer></body>", 'html'))
                 except:
                     try:
                         msg = MIMEText(ep_name + ": " + lang)
@@ -169,7 +170,7 @@ class EmailNotifier(Notifiers):
                                   msg):
                     sickrage.app.log.debug("Download notification sent to [%s] for '%s'" % (to, ep_name))
                 else:
-                    sickrage.app.log.error("Download notification ERROR: %s" % self.last_err)
+                    sickrage.app.log.warning("Download notification WARNING: %s" % self.last_err)
 
     def notify_version_update(self, new_version="??"):
         pass