+++ changelog.md
@@ -1,6 +1,16 @@
 # Changelog
 
-- * 39b6919 - 2018-05-17: Release v9.3.22 
+- * 4f89163 - 2018-05-19: Release v9.3.23 
+- * 1ff0bc5 - 2018-05-19: Converted for loop to while loop for database upgrade function 
+- * 8a40c41 - 2018-05-19: Pre-Release v9.3.23.dev2 
+- * b80633c - 2018-05-19: Added database version upgrade functionality 
+- * 249b6b0 - 2018-05-19: Converted archive_firstmatch to skip_downloaded option, skips upgrading quality of downloaded episodes 
+- * c6fd81b - 2018-05-18: Pre-Release v9.3.23.dev1 
+- * 15c6a47 - 2018-05-18: Fixed issues with downloading subtitles Fix KeyError for black and white lists Fixed NoneType error for indexer api 
+- * 0f68f72 - 2018-05-18: Converted log error to warning for Discord client Refactored SabNZBd client code 
+- * a3c7441 - 2018-05-18: Misc small fixes 
+- * 5f5592c - 2018-05-18: Misc typo corrections Converted errors to warnings for email notifications 
+- * 1630902 - 2018-05-17: Release v9.3.22 
 - * 8acce0d - 2018-05-16: Pre-Release v9.3.22.dev4 
 - * 6ea67b3 - 2018-05-15: Pre-Release v9.3.22.dev3 
 - * d2f59fc - 2018-05-15: Pre-Release v9.3.22.dev2